完成初步协议连接测试

+++ src/main/java/com/syswin/temail/gateway/codec/PacketEncoder.java
@@ -31,7 +31,7 @@ public class PacketEncoder extends MessageToByteEncoder<CDTPPacket> {
     byteBuf.writeShort(packet.getCommand());
     byteBuf.writeShort(packet.getVersion());
     byte[] headerBytes = packet.getHeader().toCDTPHeader().toByteArray();
-    byteBuf.writeInt(headerBytes.length);
+    byteBuf.writeShort(headerBytes.length);
     byteBuf.writeBytes(headerBytes);
     byteBuf.writeBytes(packet.getData());
   }
